commit
비밀번호 재입력 필요.
입력된 비밀번호가 기존 비밀번호와 일치해야 함.
비밀번호 확인 후 계정 삭제.

accounts/serializers
- UserDeleteSerializer 작성, CRU와 D의 역할을 명확히 구분하기 위해서 serializer 따로 작성
-비밀번호 입력을 위한 필드를 만들고 입력된 비밀번호가 기존 비밀번호와
일치하는지 확인한후 일치하면 진행, 아니면 오류 메시지 출력

accounts/views
- 유저 정보를 받아온 후 UserDeleteSerializer로 비밀번호 검증을 받은 후
  유효하면 회원탈퇴, 실패시 해당 메시지 출력

+++ accounts/views.py
@@ -1,5 +1,5 @@
 from django.shortcuts import get_object_or_404
-from .serializers import UserSerializer, PasswordChangeSerializer
+from .serializers import UserSerializer, PasswordChangeSerializer, UserDeleteSerializer
 from .models import User
 from profiles.models import Profile
 from rest_framework import status
@@ -50,9 +50,13 @@ class UserDetailAPIView(APIView):
             
     def delete(self, request, pk): # 회원 탈퇴
         user = self.get_object(pk)
-        user.delete()
-        data = {"delete": f"user({pk}) is deleted."}
-        return Response(data, status=status.HTTP_200_OK)
+        
+        serializer = UserDeleteSerializer(data=request.data, context={'request':request})
+        if serializer.is_valid(raise_exception=True):
+            user.delete()
+            data = {"delete": f"user({pk}) is deleted."}
+            return Response(data, status=status.HTTP_200_OK)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
     
     
 # 로그인, 로그아웃
